feat(ipc): in-register IPC fastpath — register-passing SEND/RECEIVE/SEND_SYNC (closes #11)
docs/specs/ipc-fastpath.md (issue #11, design paper).  Corrects the issue's
literal "skip full TCB context saves": no per-syscall callee-saved TCB save
exists (isr_common pushes all 15 GPRs once; deferred switch saves only rsp),
so the real win is eliminating the user-buffer round trip + canary walk by
gathering/scattering the payload only in the caller's own regs[] frame.

- SEND_FAST=74 / RECV_FAST=75 / SEND_SYNC_FAST=76, MAX_SYSCALL 74->77.
  Register budget CONFIG_IPC_FAST_PAYLOAD_BYTES=48 (6 words:
  rsi/rdi/r8/r9/r10/r11, rbp excluded; rcx/rdx carry type/data_size in the
  live int $0x80 ABI).  Pointer-free by construction -> joins k_syscall_fast[]
  (12 members), canary skip via scheduler sampling (MP-3), MP-4 AC detector
  retained.  libc __syscall_fast_send/recv/send_sync extended-asm wrappers
  (x86_64, register-pinned via asm("r8")-style locals + generic "r"
  constraints), -1 fallbacks on other arches.
- MessageQueue::pop_clamped(Message&, uint32_t): atomic selection+remove
  honoring a size clamp (one lock acquisition, zero TOCTOU); oversized best
  match is NOT removed (fail-closed INV-4).  find_best_index()/remove_at()
  extracted so pop() and pop_clamped() can never drift (INV-P).
- IPC::send_sync(..., reply_max_size=0): default 0 keeps every existing
  caller byte-identical; the fast handlers pass the budget so an oversized
  reply stays queued for a later full RECEIVE (INV-5).
- 128-bit SYSCALL_FAST_MASK (__uint128_t): syscall numbers >= 64 (DEATH_*,
  PAGER_*, *_FAST) overflowed the 64-bit mask — compiler-caught defect fixed.
- Blocking identical to the full path (block_sender / sys_receive body,
  VULN-W3 deadline, sporadic bookkeeping, WEDGE invariant, H2 untouched);
  RECV_FAST wake_sender parity after a successful pop.
- Class ipc_fastpath (14 TF_KERNEL): mask membership, ABI layout, pop_clamped
  parity, oversize-reject, roundtrip, recv-oversized-stays,
  send_sync-oversized-reply, full-queue-block, empty-block, send_sync
  roundtrip, authority, no-user-deref, latency, hybrid queue.
- Registry: all 1072->1086, MAX_TESTS 1094->1108.
- SIL 3 APPROVED (audits/report-2026-09-04T14-37-00Z.md; no S1/S2, 6 S3
  fixed: doc drift, canary doc intent, kernel-hlt deviation, arch note,
  benign transient, cosmetic).

Gates: build Errors 0; ipc_fastpath 14/14; 10 regression classes; debug all
1071/1071 (registered 1086); release all 85/85; style Errors 0.

+++ src/kernel/syscall/syscall.cpp
@@ -142,7 +142,8 @@ int syscall_path_open(const char *path, uint64_t flags) {
     // user memory, so there is nothing for the canary to protect.  FULL-path
     // syscalls keep the canary verify (see canary_check_on_full_path below).
     // Bounds are already checked above — exactly one check before indexing.
-    if (s_fastpath_enabled_ && (SYSCALL_FAST_MASK & (1ULL << number)))
+    if (s_fastpath_enabled_ &&
+        (SYSCALL_FAST_MASK & ((__uint128_t)1 << number)))
         return syscall_table_[number](arg0, arg1, arg2, arg3, regs);
 
     // FULL path: MP-3 canary verify of user segments (page-table reads, no
